% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{categorical}
\alias{categorical}
\title{Synthetic example data set for classification model.}
\format{A data frame with 100 rows and 21 columns.}
\usage{
categorical
}
\description{
Synthetic example data set for classification model.
}
\keyword{datasets}
