% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.grandforest}
\alias{importance.grandforest}
\alias{importance}
\title{Grand Forest variable importance}
\usage{
\method{importance}{grandforest}(x, ...)
}
\arguments{
\item{x}{grandforest object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Variable importance measures.
}
\description{
Extract variable importance of grandforest object.
}
\seealso{
\code{\link{grandforest}}
}
\author{
Marvin N. Wright
}
