% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network}
\alias{network}
\title{Small 20-gene example network}
\format{A data frame with 100 rows and 2 columns.}
\usage{
network
}
\description{
Each row corresponds to an edge (interaction) between the features (genes) in each column.
}
\keyword{datasets}
