% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions.grandforest.prediction}
\alias{predictions.grandforest.prediction}
\alias{predictions}
\title{Grand Forest predictions}
\usage{
\method{predictions}{grandforest.prediction}(x, ...)
}
\arguments{
\item{x}{Grand Forest prediction object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Predictions: Classes for Classification forests, Numerical values for Regressions forests and the estimated survival functions for all individuals for Survival forests.
}
\description{
Extract predictions of Grand Forest prediction object.
}
\seealso{
\code{\link{grandforest}}
}
\author{
Marvin N. Wright
}
