% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions.grandforest}
\alias{predictions.grandforest}
\title{Grand Forest predictions}
\usage{
\method{predictions}{grandforest}(x, ...)
}
\arguments{
\item{x}{grandforest object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Predictions: Classes for Classification forests, Numerical values for Regressions forests and the estimated survival functions for all individuals for Survival forests.
}
\description{
Extract training data predictions of Grand Forest object.
}
\seealso{
\code{\link{grandforest}}
}
\author{
Marvin N. Wright
}
