% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.grandforest.forest}
\alias{print.grandforest.forest}
\title{Print Grand Forest forest}
\usage{
\method{print}{grandforest.forest}(x, ...)
}
\arguments{
\item{x}{Object of class 'grandforest.forest'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of Grand Forest forest object.
}
\author{
Marvin N. Wright
}
