% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.grandforest.prediction}
\alias{print.grandforest.prediction}
\title{Print Grand Forest prediction}
\usage{
\method{print}{grandforest.prediction}(x, ...)
}
\arguments{
\item{x}{Object of class 'grandforest.prediction'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of Greand Forest prediction object.
}
\author{
Marvin N. Wright
}
