% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.r
\name{proximity.grandforest.prediction}
\alias{proximity.grandforest.prediction}
\title{Grand Forest sample proximity}
\usage{
\method{proximity}{grandforest.prediction}(x)
}
\arguments{
\item{x}{grandforest.prediction object.}
}
\value{
"dist" object
}
\description{
Extract sample proximities from grandforest.prediction object.
}
\details{
The prediction object must have been called with type = "terminalNodes".
}
