% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival}
\alias{survival}
\title{Synthetic example data set for survival model.}
\format{A data frame with 100 rows and 22 columns.}
\usage{
survival
}
\description{
Synthetic example data set for survival model.
}
\keyword{datasets}
