% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{timepoints.grandforest}
\alias{timepoints.grandforest}
\alias{timepoints}
\title{Grand Forest timepoints}
\usage{
\method{timepoints}{grandforest}(x, ...)
}
\arguments{
\item{x}{Grand Forest Survival forest object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Unique death times
}
\description{
Extract unique death times of Grand Forest Survival forest
}
\seealso{
\code{\link{grandforest}}
}
\author{
Marvin N. Wright
}
